+++ src/agent/debugger/src/lib.rs
@@ -7,3 +7,4 @@ pub mod dbghelp;
 pub mod debug_event;
 pub mod debugger;
 pub mod stack;
+mod target;
